LPD-70691 Reuse getCountries to also use when editing an object field
