+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/object/definitions/display/context/ObjectDefinitionsFieldsDisplayContext.java
@@ -12,6 +12,7 @@ import com.liferay.frontend.taglib.clay.servlet.taglib.util.CreationMenu;
 import com.liferay.list.type.service.ListTypeDefinitionService;
 import com.liferay.object.admin.rest.dto.v1_0.util.ObjectFieldUtil;
 import com.liferay.object.constants.ObjectFieldConstants;
+import com.liferay.object.field.business.type.ObjectFieldBusinessType;
 import com.liferay.object.field.business.type.ObjectFieldBusinessTypeRegistry;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectField;
@@ -65,6 +66,17 @@ public class ObjectDefinitionsFieldsDisplayContext
 		_depotEntryLocalService = depotEntryLocalService;
 		_listTypeDefinitionService = listTypeDefinitionService;
 		_objectFieldBusinessTypeRegistry = objectFieldBusinessTypeRegistry;
+	}
+
+	public List<Map<String, String>> getCountries() {
+		ObjectFieldBusinessType phoneNumberObjectFieldBusinessType =
+			_objectFieldBusinessTypeRegistry.getObjectFieldBusinessType(
+				ObjectFieldConstants.BUSINESS_TYPE_PHONE_NUMBER);
+
+		Map<String, Object> renderingProperties =
+			phoneNumberObjectFieldBusinessType.getRenderingProperties();
+
+		return (List<Map<String, String>>)renderingProperties.get("countries");
 	}
 
 	public CreationMenu getCreationMenu(ObjectDefinition objectDefinition)
