+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/object_definitions/edit_object_field.jsp
@@ -27,6 +27,8 @@ ObjectField objectField = (ObjectField)request.getAttribute(ObjectWebKeys.OBJECT
 		).put(
 			"ckEditor5Config", objectDefinitionsFieldsDisplayContext.getEditorConfig()
 		).put(
+			"countries", objectDefinitionsFieldsDisplayContext.getCountries()
+		).put(
 			"creationLanguageId", objectDefinition.getDefaultLanguageId()
 		).put(
 			"decimalSeparator", String.valueOf(decimalFormatSymbols.getDecimalSeparator())
